+++ R/openSTARS.R
@@ -114,6 +114,8 @@ if(getRversion() >= "2.15.1")
 #' South-West Wake county National Elevation Data 30m.
 #' The data has been taken from the GRASS GIS North Carolina data set.
 #' Source \url{https://grass.osgeo.org/download/sample-data/}.
+#' @section  lakes.shp:
+#' Artificial lakes.
 #' @section sites_nc.shp:
 #' Arbitrary sites along rivers in North Carolina.
 #' @section landuse.shp